projects/requirements.txt
streamlit
google-generativeai
python-dotenv
matplotlib
diffusers
torch
pillow
sqlite3
pandas
numpy
matplotlib
plotly
scikit-learn
tensorflow
keras
Pillow
transformers
spacy
requests
beautifulsoup4
watchdog
sqlalchemy
youtube-transcript-api
gensim
nltk
googlesearch-python==1.2.3